requirements/base.txt
Django==1.11.2
django-ckeditor==5.2.2
django-model-utils==3.0.0
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
psycopg2==2.7.1
pytz==2017.2
Unipath==1.1
